Improvent/home loading recyclerview

+++ app/src/main/java/com/unlam/edu/ar/videotecamoviltp/domain/model/MovieDetailsModel.kt
@@ -27,5 +27,5 @@ data class MovieDetailsModel(
     val budget: Int,
 
     @SerializedName("revenue")
-    val revenue: Int
+    val revenue: Long
 )
